refactoring. initial message. wait for resize to show tui.
